pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "mozilla-django-oidc>=4.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "celery>=5.5.3",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "pypandoc>=1.16.2",
    "bs4>=0.0.2",
    "drf-nested-routers>=0.95.0",
    "django-filter>=25.2",
    "numpy>=2.4.1",
    "scipy>=1.17.0",
    "pyjwt>=2.11.0",
    "drf-spectacular>=0.29.0",
    "python-docx>=1.2.0",
    "python-docx>=1.2.0",
    "locust>=2.43.4",
    "gunicorn>=23.0.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "locust>=2.43.4"
]
prod = [
    "gunicorn>=23.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "MacFAST.test_settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "PTH", # flake8-use-pathlib
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B024",  # abstract base class without abstract method
    "E402",  # Module level import not at top of file
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "single"
